Debloat Alpha installer and finish cleanup
