+++ .gitattributes
@@ -1,0 +1,2 @@
+# Preserve the upstream license text byte-for-byte.
+assets/fonts/share-tech-mono/OFL.txt -whitespace
