+++ assets/fonts/share-tech-mono/README.md
@@ -1,0 +1,9 @@
+# Share Tech Mono asset
+
+`ShareTechMono-Regular.ttf` and `OFL.txt` come from Google Fonts commit
+[`ec626514f79f831f1ab848a82114a0ce7e2d6372`](https://github.com/google/fonts/tree/ec626514f79f831f1ab848a82114a0ce7e2d6372/ofl/sharetechmono).
+
+- `ShareTechMono-Regular.ttf`: `sha256:9ceab1f87414829af259c0f537573ae03ef7dd3147c0b27a36a1a0beb6732677`
+- `OFL.txt`: `sha256:9d96f445b6e9c701428811d0177f894874f8d6f07ecc30d568c506542368f3ff`
+
+The font is redistributed under the SIL Open Font License included alongside it.
