+++ config/hypr/hyprlock.conf
@@ -56,7 +56,7 @@ input-field {
     outer_color       = rgba(cc1515ff)
     inner_color       = rgba(090b0ff2)
     font_color        = rgba(e7e9ebff)
-    font_family       = JetBrains Mono
+    font_family       = Iosevka
 
     dots_size         = 0.25
     dots_spacing      = 0.20
@@ -75,7 +75,7 @@ label {
     text          = $USER
     color         = rgba(e7e9ebff)
     font_size     = 13
-    font_family   = JetBrains Mono
+    font_family   = Iosevka
     position      = 44, -222
     halign        = left
     valign        = top
@@ -87,7 +87,7 @@ label {
     text          = SECURE ACCESS // PAM
     color         = rgba(646a72ff)
     font_size     = 9
-    font_family   = JetBrains Mono
+    font_family   = Iosevka
     position      = 44, -328
     halign        = left
     valign        = top
@@ -99,7 +99,7 @@ label {
     text          = $TIME
     color         = rgba(e7e9ebff)
     font_size     = 40
-    font_family   = JetBrains Mono
+    font_family   = Iosevka
     position      = 44, -116
     halign        = left
     valign        = top
@@ -111,7 +111,7 @@ label {
     text          = cmd[update:60000] date '+%Y-%m-%d  //  %A'
     color         = rgba(646a72ff)
     font_size     = 9
-    font_family   = JetBrains Mono
+    font_family   = Iosevka
     position      = 44, -190
     halign        = left
     valign        = top
@@ -123,7 +123,7 @@ label {
     text          = TSUGUMORI // SECURE SESSION
     color         = rgba(e7e9ebee)
     font_size     = 11
-    font_family   = JetBrains Mono
+    font_family   = Iosevka
     position      = 44, -44
     halign        = left
     valign        = top
@@ -135,7 +135,7 @@ label {
     text          = EMERGENCY FALLBACK CLIENT
     color         = rgba(cc1515ff)
     font_size     = 9
-    font_family   = JetBrains Mono
+    font_family   = Iosevka
     position      = 44, -76
     halign        = left
     valign        = top
@@ -147,7 +147,7 @@ label {
     text          = ARCH LINUX // HYPRLAND // HYPRLOCK
     color         = rgba(646a72dd)
     font_size     = 9
-    font_family   = JetBrains Mono
+    font_family   = Iosevka
     position      = 44, 42
     halign        = left
     valign        = bottom
@@ -159,7 +159,7 @@ label {
     text          = SESSION LOCKED // $USER
     color         = rgba(646a72dd)
     font_size     = 9
-    font_family   = JetBrains Mono
+    font_family   = Iosevka
     position      = -44, 42
     halign        = right
     valign        = bottom
