+++ config/kitty/kitty.conf
@@ -1,6 +1,6 @@
 # ── Tsugumori Terminal — Kitty Config ──
 
-# Police
+# Font
 font_family      Share Tech Mono
 bold_font        auto
 italic_font      auto
