+++ config/quickshell/widgets/lockscreen.qml
@@ -364,7 +364,7 @@ ShellRoot {
                             Text { text: "SESSION LOCKED"; font.family: "Share Tech Mono"; font.pixelSize: 9; font.letterSpacing: 2; color: "#e8e8e8" }
                         }
                         Text { text: "NODE · " + root.currentUser + "@arch"; font.family: "Share Tech Mono"; font.pixelSize: 9; font.letterSpacing: 2; color: "#e8e8e8" }
-                        Text { text: "SESSION LOCK ACTIVE"; font.family: "Share Tech Mono"; font.pixelSize: 8; color: "#e8e8e8"; opacity: 0.7 }
+                        Text { text: "セッションロック中"; font.family: "Noto Sans CJK JP"; font.pixelSize: 8; color: "#e8e8e8"; opacity: 0.7 }
                     }
                 }
                 Item {
@@ -490,7 +490,7 @@ ShellRoot {
                             }
                             Text { text: root.currentUser.toUpperCase(); font.family: "Share Tech Mono"; font.pixelSize: 13; font.letterSpacing: 3; color: "#e8e8e8"; anchors.horizontalCenter: parent.horizontalCenter }
                             Item { width: 1; height: 4 }
-                            Text { text: "UNIT · ACTIVE"; font.family: "Share Tech Mono"; font.pixelSize: 8; color: "#7a7358"; anchors.horizontalCenter: parent.horizontalCenter }
+                            Text { text: "ユニット · アクティブ"; font.family: "Noto Sans CJK JP"; font.pixelSize: 8; color: "#7a7358"; anchors.horizontalCenter: parent.horizontalCenter }
                             Item { width: 1; height: 20 }
                             Text { text: root.clockStr; font.family: "Share Tech Mono"; font.pixelSize: 46; font.letterSpacing: 2; color: "#e8e8e8"; anchors.horizontalCenter: parent.horizontalCenter }
                             Item { width: 1; height: 6 }
@@ -523,7 +523,7 @@ ShellRoot {
                                         Keys.onReturnPressed: root.doAuth()
                                         Keys.onEscapePressed: root.lockInput = ""
                                         Text { visible: parent.text === ""; anchors.verticalCenter: parent.verticalCenter
-                        text: root.lockPending ? "authenticating..." : "password..."; font.family: "Share Tech Mono"; font.pixelSize: 11; font.italic: true; color: "#7a7358"; opacity: 0.5
+                                            text: root.lockPending ? "authenticating..." : "password..."; font.family: "Share Tech Mono"; font.pixelSize: 11; font.italic: true; color: "#7a7358"; opacity: 0.5
                                         }
                                     }
                                 }
